Add compiled executable and object files for the project
- Added outDebug.exe for the project build.
- Added sistemaDeMatriculaEscolar.o object file for the project build.

+++ projetoMatricula/sistemaDeMatriculaEscolar.c
@@ -239,5 +239,5 @@ float calculaMediaTurma(Aluno turma[], int n){
         soma += calculaMedia(turma, i);
     }
 
-    return soma/n;
+    return soma / (float)n;
 }
